Rails: Load rake tasks in the engine
By adding the rake tasks in the engine, it makes Rails load them
properly so the rails CLI can pick it up.

With these changes, you can do `rails vapey:reindex` and it will run
this task.

If you run `rails` alone, the task will also be listed like all the
others.

I changed the directory of the tasks as tasks that are required
at runtime are usually located under lib/gem_name.

Tasks that are used for development purposes or that want to be excluded
from the bundled gem are usually set under lib/tasks.